+++ drivers/misc/Kconfig
@@ -5,6 +5,7 @@
 
 menu "Miscellaneous Drivers"
 
+source "drivers/misc/ethos_u/Kconfig"
 source "drivers/misc/ft8xx/Kconfig"
 source "drivers/misc/grove_lcd_rgb/Kconfig"
 source "drivers/misc/pio_rpi_pico/Kconfig"
