+++ drivers/misc/ethos_u/ethos_u_common.h
@@ -1,0 +1,31 @@
+/*
+ * SPDX-FileCopyrightText: <text>Copyright 2021-2022, 2024 Arm Limited and/or its
+ * affiliates <open-source-office@arm.com></text>
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#ifndef ZEPHYR_DRIVERS_MISC_ETHOS_U_ETHOS_U_COMMON_H_
+#define ZEPHYR_DRIVERS_MISC_ETHOS_U_ETHOS_U_COMMON_H_
+
+#include <ethosu_driver.h>
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+struct ethosu_dts_info {
+	void *base_addr;
+	bool secure_enable;
+	bool privilege_enable;
+	void (*irq_config)(void);
+};
+
+struct ethosu_data {
+	struct ethosu_driver drv;
+};
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif /* ZEPHYR_DRIVERS_MISC_ETHOS_U_ETHOS_U_COMMON_H_ */
